requirements.txt
requests
django
django-cors-headers
djangorestframework
google-generativeai
g4f[all]